+++ src/store/index.ts
@@ -1,6 +1,9 @@
 import { ref } from "vue";
 import { defineStore } from "pinia";
 
+import { OrderInfoData } from "@/models/orderInfo";
+import moment from "moment";
+
 export const useGeneralStore = defineStore("generalStore", () => {
   const date = ref(new Date());
   const orderEvent = ref(0);
@@ -33,6 +36,21 @@ export const useGeneralStore = defineStore("generalStore", () => {
     isWindowActive.value = flag;
   };
 
+  const orderObj = ref<{ [key: string]: OrderInfoData[] }>({});
+  const setOrders = (orders: OrderInfoData[]) => {
+    orderObj.value = orders.reduce(
+      (tmp: { [key: string]: OrderInfoData[] }, order: OrderInfoData) => {
+        const day = moment(order.timePlaced.toDate()).format("YYYY-MM-DD");
+        if (!tmp[day]) {
+          tmp[day] = [];
+        }
+        tmp[day].push(order);
+        return tmp;
+      },
+      {},
+    );
+  };
+
   return {
     date,
     updateDate,
@@ -51,5 +69,8 @@ export const useGeneralStore = defineStore("generalStore", () => {
 
     isWindowActive,
     setActive,
+
+    orderObj,
+    setOrders,
   };
 });
